crypto: keystone: add low level crypto hardware interface code
This patch adds the driver support for keystone SA in sideband mode.

The driver registers algorithm implementations in the Kernel crypto
framework. Since the primary use case in kernel is to enable HW offload
of IPSec ESP crypto operations, the driver is currently supporting only
authenticated encryption (AEAD) mode. Following algorithms are currently
supported:

 - authenc(hmac(sha1),cbc(aes))
 - authenc(hmac(sha1),cbc(des3_ede))
 - authenc(hmac(sha1),ecb(cipher_null))

The driver is based on TI SA Low Level Driver (LLD).
In order to simplify maintainability and easy porting of future LLD
version the kernel driver code is kept as much close to LLD as possible.

To make the driver more readable it was split into several commits.
This is the first commit which introduces the header files and
keystone-sa-lld.c. The keystone-sa-lld.c re-implements/copies number of
helper functions provided by LLD.

Signed-off-by: Sandeep Nair <sandeep_n@ti.com>
Signed-off-by: Tinku Mannan <tmannan@ti.com>
Signed-off-by: Hao Zhang <hzhang@ti.com>
Signed-off-by: Vitaly Andrianov <vitalya@ti.com>