net: netcp: support of multiple subqueues per interface
This patch adds the support of multiple subqueues per
interface.  This is particularly useful when support
of NETCP QoS is enabled.  In that case, each subqueue
of the interface will be mapped to support one class
of service.

In this implementation, one subqueue is mapped to a
NETCP txpipe with its own tx-pool and tx-threshold
and rx-threshold.  This allows the flexibility of
a configurable number of descriptors/buffers in order
to serve the need of that particular class of service.

This patch is backward compatible to the support of
one subqueue configured in the way prior to this patch.
But it won't be able to support the up coming NETCP
QoS with the legacy subqueue configurations.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>